NS-757: Decouples SIS terms from the rest of SIS data
