+++ config/default.py
@@ -197,6 +197,7 @@ RDS_SCHEMA_E_I = 'boac_advising_e_i'
 RDS_SCHEMA_METADATA = 'metadata'
 RDS_SCHEMA_SIS_ADVISING_NOTES = 'sis_advising_notes'
 RDS_SCHEMA_SIS_INTERNAL = 'sis_data'
+RDS_SCHEMA_SIS_TERMS = 'sis_terms'
 RDS_SCHEMA_STUDENT = 'student'
 RDS_SCHEMA_UNDERGRADS = 'undergrads'
 
@@ -242,6 +243,7 @@ REDSHIFT_SCHEMA_GRADESCOPE = 'External gradescope schema name'
 REDSHIFT_SCHEMA_LRS = 'External LRS schema name'
 REDSHIFT_SCHEMA_SIS = 'External SIS schema name'
 REDSHIFT_SCHEMA_SIS_ADVISING_NOTES = 'External SIS Advising Notes schema name'
+REDSHIFT_SCHEMA_SIS_TERMS = 'External SIS Terms schema name'
 REDSHIFT_SCHEMA_UNDERGRADS_EXTERNAL = 'External Undergrads schema name'
 
 STUDENT_API_ID = 'secretid'
