+++ config/test.py
@@ -39,6 +39,7 @@ RDS_SCHEMA_COE = 'coe_test'
 RDS_SCHEMA_E_I = 'e_i_test'
 RDS_SCHEMA_METADATA = 'nessie_metadata_test'
 RDS_SCHEMA_SIS_INTERNAL = 'sis_internal_test'
+RDS_SCHEMA_SIS_TERMS = 'sis_terms_test'
 RDS_SCHEMA_STUDENT = 'student_test'
 
 REDSHIFT_DATABASE = 'nessie_redshift_test'
@@ -68,6 +69,7 @@ REDSHIFT_SCHEMA_SIS = 'sis_test'
 REDSHIFT_SCHEMA_SIS_ADVISING_NOTES = 'sis_advising_notes_external_test'
 REDSHIFT_SCHEMA_SIS_ADVISING_NOTES_INTERNAL = 'sis_advising_notes_test'
 REDSHIFT_SCHEMA_SIS_INTERNAL = 'sis_internal_test'
+REDSHIFT_SCHEMA_SIS_TERMS = 'sis_terms_test'
 REDSHIFT_SCHEMA_STUDENT = 'student_test'
 REDSHIFT_SCHEMA_UNDERGRADS = 'undergrads_test'
 REDSHIFT_SCHEMA_UNDERGRADS_EXTERNAL = 'undergrads_external_test'
