+++ nessie/jobs/scheduling.py
@@ -220,6 +220,7 @@ def run_startup_jobs(_app):
     from nessie.jobs.create_asc_schema import CreateAscSchema
     from nessie.jobs.create_metadata_schema import CreateMetadataSchema
     from nessie.jobs.create_rds_indexes import CreateRdsIndexes
+    from nessie.jobs.create_sis_terms_schema import CreateSisTermsSchema
     from nessie.jobs.create_student_schema import CreateStudentSchema
 
     if _app.config['JOB_SCHEDULING_ENABLED'] and os.environ.get('NESSIE_ENV') != 'test':
@@ -228,6 +229,7 @@ def run_startup_jobs(_app):
         CreateAscSchema().run()
         CreateMetadataSchema().run()
         CreateRdsIndexes().run()
+        CreateSisTermsSchema().run()
         CreateStudentSchema().run()
