+++ nessie/lib/util.py
@@ -150,6 +150,7 @@ def resolve_sql_template_string(template_string, **kwargs):
         'rds_schema_metadata': app.config['RDS_SCHEMA_METADATA'],
         'rds_schema_sis_advising_notes': app.config['RDS_SCHEMA_SIS_ADVISING_NOTES'],
         'rds_schema_sis_internal': app.config['RDS_SCHEMA_SIS_INTERNAL'],
+        'rds_schema_sis_terms': app.config['RDS_SCHEMA_SIS_TERMS'],
         'rds_schema_student': app.config['RDS_SCHEMA_STUDENT'],
         'rds_schema_undergrads': app.config['RDS_SCHEMA_UNDERGRADS'],
         'redshift_app_boa_user': app.config['REDSHIFT_APP_BOA_USER'],
@@ -176,6 +177,7 @@ def resolve_sql_template_string(template_string, **kwargs):
         'redshift_schema_sis_advising_notes': app.config['REDSHIFT_SCHEMA_SIS_ADVISING_NOTES'],
         'redshift_schema_sis_advising_notes_internal': app.config['REDSHIFT_SCHEMA_SIS_ADVISING_NOTES_INTERNAL'],
         'redshift_schema_sis_internal': app.config['REDSHIFT_SCHEMA_SIS_INTERNAL'],
+        'redshift_schema_sis_terms': app.config['REDSHIFT_SCHEMA_SIS_TERMS'],
         'redshift_schema_student': app.config['REDSHIFT_SCHEMA_STUDENT'],
         'redshift_schema_undergrads': app.config['REDSHIFT_SCHEMA_UNDERGRADS'],
         'redshift_schema_undergrads_external': app.config['REDSHIFT_SCHEMA_UNDERGRADS_EXTERNAL'],
