+++ nessie/sql_templates/restore_rds_user_privileges.template.sql
@@ -29,14 +29,17 @@ GRANT USAGE ON SCHEMA {rds_schema_asc} TO {rds_app_boa_user};
 GRANT USAGE ON SCHEMA {rds_schema_coe} TO {rds_app_boa_user};
 GRANT USAGE ON SCHEMA {rds_schema_sis_advising_notes} TO {rds_app_boa_user};
 GRANT USAGE ON SCHEMA {rds_schema_sis_internal} TO {rds_app_boa_user};
+GRANT USAGE ON SCHEMA {rds_schema_sis_terms} TO {rds_app_boa_user};
 GRANT USAGE ON SCHEMA {rds_schema_student} TO {rds_app_boa_user};
 GRANT SELECT ON ALL TABLES IN SCHEMA {rds_schema_asc} TO {rds_app_boa_user};
 GRANT SELECT ON ALL TABLES IN SCHEMA {rds_schema_coe} TO {rds_app_boa_user};
 GRANT SELECT ON ALL TABLES IN SCHEMA {rds_schema_sis_advising_notes} TO {rds_app_boa_user};
 GRANT SELECT ON ALL TABLES IN SCHEMA {rds_schema_sis_internal} TO {rds_app_boa_user};
+GRANT SELECT ON ALL TABLES IN SCHEMA {rds_schema_sis_terms} TO {rds_app_boa_user};
 GRANT SELECT ON ALL TABLES IN SCHEMA {rds_schema_student} TO {rds_app_boa_user};
 ALTER DEFAULT PRIVILEGES IN SCHEMA {rds_schema_asc} GRANT SELECT ON TABLES TO {rds_app_boa_user};
 ALTER DEFAULT PRIVILEGES IN SCHEMA {rds_schema_coe} GRANT SELECT ON TABLES TO {rds_app_boa_user};
 ALTER DEFAULT PRIVILEGES IN SCHEMA {rds_schema_sis_advising_notes} GRANT SELECT ON TABLES TO {rds_app_boa_user};
 ALTER DEFAULT PRIVILEGES IN SCHEMA {rds_schema_sis_internal} GRANT SELECT ON TABLES TO {rds_app_boa_user};
+ALTER DEFAULT PRIVILEGES IN SCHEMA {rds_schema_sis_terms} GRANT SELECT ON TABLES TO {rds_app_boa_user};
 ALTER DEFAULT PRIVILEGES IN SCHEMA {rds_schema_student} GRANT SELECT ON TABLES TO {rds_app_boa_user};
