+++ tests/test_jobs/test_create_sis_schema.py
@@ -26,35 +26,11 @@ ENHANCEMENTS, OR MODIFICATIONS.
 from datetime import datetime, timedelta
 import json
 
-import mock
-from nessie.externals import rds, s3
+from nessie.externals import s3
 from nessie.jobs.background_job import BackgroundJobError
-from nessie.lib.util import get_s3_sis_daily_path, resolve_sql_template
+from nessie.lib.util import get_s3_sis_daily_path
 import pytest
 from tests.util import capture_app_logs, mock_s3
-
-
-@pytest.fixture
-def term_definitions(app):
-    rds_schema = app.config['RDS_SCHEMA_SIS_INTERNAL']
-    rds.execute(f'DROP SCHEMA {rds_schema} CASCADE')
-    rds.execute(resolve_sql_template('create_rds_indexes.template.sql'))
-    rds.execute(f"""INSERT INTO {rds_schema}.term_definitions
-        (term_id, term_name, term_begins, term_ends)
-        VALUES
-        ('2172', 'Spring 2017', '2017-01-10', '2017-05-12'),
-        ('2175', 'Summer 2017', '2017-05-22', '2017-08-11'),
-        ('2178', 'Fall 2017', '2017-08-16', '2017-12-15'),
-        ('2182', 'Spring 2018', '2018-01-09', '2018-05-11'),
-        ('2185', 'Summer 2018', '2018-05-21', '2018-08-10'),
-        ('2188', 'Fall 2018', '2018-08-15', '2018-12-14'),
-        ('2192', 'Spring 2019', '2019-01-15', '2019-05-17'),
-        ('2195', 'Summer 2019', '2019-05-28', '2019-08-16'),
-        ('2198', 'Fall 2019', '2019-08-21', '2019-12-20'),
-        ('2202', 'Spring 2020', '2020-01-14', '2020-05-15'),
-        ('2205', 'Summer 2020', '2020-05-26', '2020-08-14'),
-        ('2208', 'Fall 2020', '2020-08-19', '2020-12-18')
-    """)
 
 
 class TestCreateSisSchema:
@@ -91,18 +67,6 @@ class TestCreateSisSchema:
                 with pytest.raises(BackgroundJobError) as e:
                     CreateSisSchema().update_manifests()
                 assert 'Expected filename enrollments-2178.gz not found in S3, aborting' in str(e.value)
-
-    @mock.patch('nessie.jobs.create_sis_schema.datetime', autospec=True)
-    def test_refresh_current_term_index(self, mock_datetime, app, term_definitions):
-        mock_datetime.now.return_value = datetime(year=2018, month=5, day=1, hour=5, minute=21)
-        from nessie.jobs.create_sis_schema import CreateSisSchema
-        CreateSisSchema().refresh_current_term_index()
-
-        rds_schema = app.config['RDS_SCHEMA_SIS_INTERNAL']
-        rows = rds.fetch(f'SELECT * FROM {rds_schema}.current_term_index')
-        assert len(rows) == 1
-        assert rows[0]['current_term_name'] == 'Spring 2018'
-        assert rows[0]['future_term_name'] == 'Fall 2018'
 
     def _upload_data_to_s3(self, daily_path, historical_path):
         s3.upload_data('some futuristic course data', f'{daily_path}/courses/courses-2182.gz')
